# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Soulful Corporate Wellbeing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/company:
    get:
      operationId: listCompanySubscriptions
      tags: [subscriptions]
      summary: List active corporate subscriptions
      responses:
        "200":
          description: Corporate subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanySubscription"
    post:
      operationId: createCompanySubscription
      tags: [subscriptions]
      summary: Subscribe a company to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanySubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySubscription"
  /subscriptions/practitioner:
    get:
      operationId: listPractitionerSubscriptions
      tags: [subscriptions]
      summary: List active practitioner subscriptions
      responses:
        "200":
          description: Practitioner subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PractitionerSubscription"
    post:
      operationId: createPractitionerSubscription
      tags: [subscriptions]
      summary: Subscribe a practitioner to a listing plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PractitionerSubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PractitionerSubscription"

  # ── Stripe payments ─────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    CompanySubscription:
      type: object
      required: [id, companyId, planId, status, startDate, companyName, planName, priceGbp]
      properties:
        id:          { type: integer }
        companyId:   { type: integer }
        planId:      { type: integer }
        status:      { type: string, enum: [active, cancelled, past_due] }
        startDate:   { type: string }
        endDate:     { type: ["string", "null"] }
        companyName: { type: string }
        planName:    { type: string }
        priceGbp:    { type: number }
    CompanySubscriptionInput:
      type: object
      required: [companyId, planId]
      properties:
        companyId: { type: integer }
        planId:    { type: integer }
    PractitionerSubscription:
      type: object
      required: [id, practitionerId, planId, status, startDate, practitionerName, planName, priceGbp]
      properties:
        id:               { type: integer }
        practitionerId:   { type: integer }
        planId:           { type: integer }
        status:           { type: string, enum: [active, cancelled, past_due] }
        startDate:        { type: string }
        endDate:          { type: ["string", "null"] }
        practitionerName: { type: string }
        planName:         { type: string }
        priceGbp:         { type: number }
    PractitionerSubscriptionInput:
      type: object
      required: [practitionerId, planId]
      properties:
        practitionerId: { type: integer }
        planId:         { type: integer }

    # ── Reviews ──────────────────────────────────────────────────────────────
...